update: one decision for whether the OS bundle is already running
The OS bundle is the rootfs plus the boot FIT (kernel + initramfs), but
"already up to date" was decided from the rootfs id alone, in three places:
the download step skipped fetching the bundle, the already-current check
reported nothing to do, and activation skipped applying it. A runtime whose
only difference is the initramfs therefore went active with its bundle never
even downloaded and the device kept booting the old initrd - exactly what
happened on imx8mp-evk (manifest present, cf136cbe.raw absent, no reboot).

os_update::os_bundle_satisfied() is now the single predicate: rootfs id
against the running os-release, initramfs id against the active runtime's
record (the running system keeps none of its own). All three sites use it.
